+++ packages/pwa-kit-dev/bin/pwa-kit-dev.js
@@ -14,6 +14,7 @@ const program = require('commander')
 const validator = require('validator')
 const {execSync: _execSync} = require('child_process')
 const {getConfig} = require('@salesforce/pwa-kit-runtime/utils/ssr-config')
+const openInBrowser = require('open')
 
 // Scripts in ./bin have never gone through babel, so we
 // don't have a good pattern for mixing compiled/un-compiled
@@ -29,6 +30,9 @@ const scriptUtils = (() => {
         return require('../utils/script-utils')
     }
 })()
+
+/** Environment variable used by the start/push command --open flag. */
+const PWA_KIT_OPEN_URL = 'PWA_KIT_OPEN_URL'
 
 const colors = {
     warn: 'yellow',
@@ -134,9 +138,13 @@ const main = async () => {
         return program
             .command(name)
             .addOption(
-                new program.Option('--cloud-origin <origin>', 'the API origin to connect to')
+                new program.Option(
+                    '--cloud-origin <origin>',
+                    '[internal] the API origin to connect to'
+                )
                     .default(scriptUtils.DEFAULT_CLOUD_ORIGIN)
                     .env('CLOUD_API_BASE')
+                    .hideHelp()
             )
             .addOption(
                 new program.Option(
@@ -237,7 +245,15 @@ const main = async () => {
                 'args to pass through to babel-node'
             ).default('--extensions ".js,.jsx,.ts,.tsx"')
         )
-        .action(async ({inspect, noHMR, babelArgs}) => {
+        .addOption(
+            new program.Option('-o, --open <url>', 'initial URL to load after the server starts')
+                .default('/')
+                .env(PWA_KIT_OPEN_URL)
+        )
+        .addOption(
+            new program.Option('--no-open', 'do not launch a browser when the dev server starts')
+        )
+        .action(async ({inspect, noHMR, babelArgs, open}) => {
             // We use @babel/node instead of node because we want to support ES6 import syntax
             const babelNode = p.join(
                 require.resolve('webpack'),
@@ -257,7 +273,10 @@ const main = async () => {
             execSync(`${babelNode} ${inspect ? '--inspect' : ''} ${babelArgs} ${entrypoint}`, {
                 env: {
                     ...process.env,
-                    ...(noHMR ? {HMR: 'false'} : {})
+                    // NOTE: The logic in build-dev-server.js enforces that only URLs pointing to
+                    // the dev server are actually opened
+                    ...(open && {[PWA_KIT_OPEN_URL]: open}),
+                    ...(noHMR && {HMR: 'false'})
                 }
             })
         })
@@ -345,6 +364,11 @@ const main = async () => {
         .addOption(
             new program.Option('-w, --wait', 'wait for the deployment to complete before exiting')
         )
+        .addOption(
+            new program.Option('-o, --open <url>', 'URL to open after the deployment succeeds')
+                .env(PWA_KIT_OPEN_URL)
+                .implies({wait: true})
+        )
         .action(
             async ({
                 buildDirectory,
@@ -354,6 +378,7 @@ const main = async () => {
                 cloudOrigin,
                 credentialsFile,
                 wait,
+                open: urlToOpen,
                 user,
                 key
             }) => {
@@ -365,7 +390,7 @@ const main = async () => {
                     process.env.DEPLOY_TARGET = target
                 } else if (wait) {
                     throw new Error(
-                        'You must provide a target to deploy to when using --wait to wait for deployment to finish.'
+                        'You must provide a deployment target when using --wait to wait for deployment to finish.'
                     )
                 }
 
@@ -409,7 +434,19 @@ const main = async () => {
                 warnings.forEach(warn)
                 if (wait) {
                     success('Bundle Uploaded - waiting for deployment to complete')
-                    await client.waitForDeploy(projectSlug, target)
+                    const hostname = await client.waitForDeploy(projectSlug, target)
+                    success('Deployment completed')
+                    if (urlToOpen) {
+                        // If given a relative path, assume it's for the deployed site
+                        const url = new URL(urlToOpen, `https://${hostname}`)
+                        // Allow pretty much any URL, because users may want to open a third party
+                        // site for next steps after deploying, but don't allow weird input.
+                        if (!url.protocol !== 'http:' && url.protocol !== 'https:') {
+                            warn(`Refusing to open invalid URL: ${urlToOpen}`)
+                        } else {
+                            openInBrowser(url.href)
+                        }
+                    }
                 } else {
                     success('Bundle Uploaded')
                 }