+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## v3.5.0-dev (Jan 22, 2024)
 ## v3.4.0 (Jan 19, 2024)
 
+- Add `--open <url>` flag to `pwa-kit-dev` `start` and `push` commands. [#1592](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1592)
+  - `pwa-kit-dev start --open <url>` opens the specified URL when the dev server starts, rather than the homepage.
+  - `pwa-kit-dev start --no-open` prevents any URL from opening when the dev server starts.
+  - `pwa-kit-dev push -s <project> -t <environment> --open <url>` opens the specified URL after the deployment succeeds.
 - Add support for node 20 [#1612](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1612)
 
 ## v3.3.0 (Dec 08, 2023)